# ENH: Speedup clip for floating point