@eendebakpt 2024-04-15
Some small comments, but the implementation and performance gain look good!