@mattip 2024-04-17
We discussed this a recent triage meeting. This seems to make sense. Would you get the same speed improvement by using the `-O3` macro? If we stay with this specialization you could also remove the floating point error checks from the other loop.