@pijyoi 2024-04-17
>Would you get the same speed improvement by using the -O3 macro?

The generated code is the same for -O2 and -O3.
-O2 : https://godbolt.org/z/53crsc1hG
-O3 : https://godbolt.org/z/Ys1eMx4Ej

>If we stay with this specialization you could also remove the floating point error checks from the other loop.

I take this to mean that "long double" should also use this specialized path. I have added it and updated the asv output above according. Also added int32 and int64 benchmarks for comparison's sake.
So it turns out that "long double" which is `fp128` on Linux doesn't get any speedup on my machine.

However "complex floating" types are still using the original path and thus the floating point error checks cannot be removed.