@pijyoi 2024-04-18
I ran asv on another machine (Intel), and longdouble does get an improvement.
(The machine where it didn't have an improvement was an AMD)

| Change   | Before [c2a9a1cb] <speedup-float-clip~1>   | After [96ed0d25] <speedup-float-clip>   |   Ratio | Benchmark (Parameter)                                              |
|----------|--------------------------------------------|-----------------------------------------|---------|--------------------------------------------------------------------|
| -        | 308±0.7μs                                  | 228±0.5μs                               |    0.74 | bench_clip.ClipFloat.time_clip(<class 'numpy.longdouble'>, 100000) |
| -        | 47.5±2μs                                   | 25.7±0.5μs                              |    0.54 | bench_clip.ClipFloat.time_clip(<class 'numpy.float64'>, 100000)    |
| -        | 22.6±0.2μs                                 | 12.1±0.1μs                              |    0.53 | bench_clip.ClipFloat.time_clip(<class 'numpy.float32'>, 100000)    |