@j9ac9k 2024-04-18
I have an arm64 chip (M2 Max), I just ran the same asv benchmark as above and got the following results 

| Change   | Before [c2a9a1cb] <speedup-float-clip~1>   | After [96ed0d25] <speedup-float-clip>   |   Ratio | Benchmark (Parameter)                                              |
|----------|--------------------------------------------|-----------------------------------------|---------|--------------------------------------------------------------------|
| -        | 1.13±0.03μs                                | 1.08±0μs                                |    0.95 | bench_clip.ClipFloat.time_clip(<class 'numpy.float64'>, 100)       |
| -        | 90.1±0.9μs                                 | 60.3±0.3μs                              |    0.67 | bench_clip.ClipFloat.time_clip(<class 'numpy.float32'>, 100000)    |
| -        | 90.6±0.3μs                                 | 60.3±0.1μs                              |    0.67 | bench_clip.ClipFloat.time_clip(<class 'numpy.float64'>, 100000)    |
| -        | 92.2±2μs                                   | 60.7±0.4μs                              |    0.66 | bench_clip.ClipFloat.time_clip(<class 'numpy.longdouble'>, 100000) |