@seberg 2024-04-18
> The generated code is the same for -O2 and -O3.

The idea was to effectively compile the whole file with it to see if the compiler isn't smart enough to do those some of those optimizations then.  We also have `NPY_GCC_OPT_3` which is a function attribute.  Unfortunately, that really only works with GCC, though.

But maybe it isn't and to really use it, it seems a bit like the ufunc build needs to be split into a default part a part that defaults to `-O3` for loops where we consider it worthwhile...