@pijyoi 2024-04-18
So we can compare the following 2 snippets at -O3:
https://godbolt.org/z/f7ofbWdqr
```c++
void clip_loop1(const float *input, float *output, size_t nelems, float min_val, float max_val)
{
    for (size_t k=0; k < nelems; k++) {
        output[k] = _NPY_CLIP(input[k], min_val, max_val);
    }
}

void clip_loop2(const float *input, float *output, size_t nelems, float min_val, float max_val)
{
    for (size_t k=0; k < nelems; k++) {
        float x = input[k];
        if (x < min_val) x = min_val;
        if (x > max_val) x = max_val;
        output[k] = x;
    }
}
```

We can then compare the SIMD-dified 4 at a time loop.

main vectorized loop of `clip_loop1`
Large portions of the code is masking and selecting.
```
.L5:
        movups  xmm3, XMMWORD PTR [rcx+rax]
        movaps  xmm0, xmm6
        movaps  xmm2, xmm6
        movaps  xmm4, xmm3
        cmpltps xmm0, xmm3
        cmpordps        xmm4, xmm3
        pandn   xmm0, xmm4
        andps   xmm2, xmm0
        andnps  xmm0, xmm3
        orps    xmm0, xmm2
        movaps  xmm2, xmm0
        cmpltps xmm2, xmm5
        movdqa  xmm8, xmm2
        pandn   xmm2, xmm4
        pand    xmm8, xmm4
        movaps  xmm4, xmm3
        cmpunordps      xmm4, xmm3
        andps   xmm3, xmm4
        andnps  xmm4, xmm5
        orps    xmm4, xmm3
        movaps  xmm3, xmm5
        andps   xmm3, xmm2
        andnps  xmm2, xmm4
        orps    xmm2, xmm3
        movaps  xmm3, xmm0
        movaps  xmm0, xmm8
        andps   xmm3, xmm8
        andnps  xmm0, xmm2
        orps    xmm0, xmm3
        movups  XMMWORD PTR [rsi+rax], xmm0
        add     rax, 16
        cmp     rax, rdi
        jne     .L5
```

main vectorized loop of `clip_loop2`
In comparison, this code is simple and looks just like the non-vectorized version.
```
.L47:
        movups  xmm6, XMMWORD PTR [rcx+rax]
        movaps  xmm3, xmm5
        movaps  xmm2, xmm4
        maxps   xmm3, xmm6
        minps   xmm2, xmm3
        movups  XMMWORD PTR [rsi+rax], xmm2
        add     rax, 16
        cmp     rdi, rax
        jne     .L47
```