@pijyoi 2024-04-19
I tried annotating the 3 floating point functions like so. Based on inspection of the disassembly of the compiled object file, this does trigger the SIMD-dification. (For whatever reason, NPY_GCC_OPT_3 didn't work). 
```c++
NPY_NO_EXPORT void
__attribute__((optimize("O3")))
FLOAT_clip(char **args, npy_intp const *dimensions, npy_intp const *steps,
           void *NPY_UNUSED(func))
{
    ...
}
```

As reported by asv,
1) turning on -O3 did not improve the original timings of the main branch
2) turning on -O3 did not further improve the timings of this PR