@seberg 2024-04-19
Thanks for checking, seems like the float special path is necessary for lifting the nan handling (at least I presume that is the mechanism here).  Also nice to know that O3 doesn't actually help with the loop at all :).
Might still wonder if we can integrate that without having to duplicating the code, but not sure and I am OK with it also.