@pijyoi 2024-04-20
I got `spin bench` to run on native Windows (same AMD machine, so the timings can be compared directly against the results in the top-most post) .
As can be seen, main branch clip timings are much worse on Windows (MSVC) than on Linux (GCC); by almost a factor of 4 for `float32`.
This PR has made the `float32` and `float64` timings equivalent on both MSVC and GCC.

| Change   | Before [c2a9a1cb] <speedup-float-clip~1>   | After [96ed0d25] <speedup-float-clip>   |   Ratio | Benchmark (Parameter)                                              |
|----------|--------------------------------------------|-----------------------------------------|---------|--------------------------------------------------------------------|
| -        | 16.4±1μs                                   | 14.5±0.2μs                              |    0.88 | bench_clip.ClipInteger.time_clip(<class 'numpy.int32'>, 100000)    |
| -        | 125±3μs                                    | 29.7±0.4μs                              |    0.24 | bench_clip.ClipFloat.time_clip(<class 'numpy.float64'>, 100000)    |
| -        | 128±3μs                                    | 30.7±0.6μs                              |    0.24 | bench_clip.ClipFloat.time_clip(<class 'numpy.longdouble'>, 100000) |
| -        | 124±6μs                                    | 18.3±0.8μs                              |    0.15 | bench_clip.ClipFloat.time_clip(<class 'numpy.float32'>, 100000)    |